cleanup: trim docstrings, remove dead code, consolidate PD tests (8→5 files)
