+++ python/sgl_jax/test/disaggregation/test_pd_mixin_event_loop.py
@@ -978,3 +978,88 @@ def test_different_rids_same_transfer_id_uses_shared_id():
 
     (metadata,), _ = d_receiver.init.call_args
     assert metadata.uuid == "shared-xfer-42"
+
+
+# ---- shutdown handler (merged from test_disagg_shutdown_handler.py) ----------
+
+
+def test_disagg_shutdown_handler_unregisters_and_drains():
+    from sgl_jax.srt.managers.scheduler import _make_disagg_shutdown
+
+    scheduler = mock.MagicMock()
+    scheduler.disagg_bootstrap_key = "bkey"
+    scheduler.disagg_bootstrap_client = mock.MagicMock()
+    scheduler.disagg_heartbeat = mock.MagicMock()
+    scheduler.disagg_kv_manager = mock.MagicMock()
+
+    fn = _make_disagg_shutdown(scheduler, "prefill")
+    fn()
+    scheduler.disagg_bootstrap_client.unregister_prefill.assert_called_once_with(
+        "bkey"
+    )
+    scheduler.disagg_heartbeat.stop.assert_called_once()
+    scheduler.disagg_kv_manager.graceful_shutdown.assert_called_once()
+    scheduler.disagg_kv_manager.zmq_notifier.stop.assert_called_once()
+    # Idempotent.
+    fn()
+    scheduler.disagg_bootstrap_client.unregister_prefill.assert_called_once()
+
+
+# ---- debug utils (merged from test_pd_debug_utils.py) -----------------------
+
+
+def test_kv_debug_enabled_honors_flag_and_req_filter(monkeypatch):
+    from sgl_jax.srt.disaggregation.debug_utils import kv_debug_enabled
+
+    monkeypatch.delenv("SGL_JAX_PD_DEBUG_KV", raising=False)
+    monkeypatch.delenv("SGL_JAX_PD_DEBUG_REQ_ID", raising=False)
+    assert not kv_debug_enabled("req-1")
+
+    monkeypatch.setenv("SGL_JAX_PD_DEBUG_KV", "1")
+    assert kv_debug_enabled("req-1")
+
+    monkeypatch.setenv("SGL_JAX_PD_DEBUG_REQ_ID", "bucket16")
+    assert kv_debug_enabled("probe-bucket16-req")
+    assert not kv_debug_enabled("probe-bucket8-req")
+
+
+def test_build_kv_debug_snapshot_captures_global_and_page_digests():
+    from sgl_jax.srt.disaggregation.debug_utils import build_kv_debug_snapshot
+
+    kv = np.arange(2 * 3 * 4, dtype=np.int16).reshape(2, 3, 4)
+    snapshot = build_kv_debug_snapshot(kv)
+
+    assert snapshot.shape == (2, 3, 4)
+    assert snapshot.dtype == "int16"
+    assert len(snapshot.global_digest) == 16
+    assert len(snapshot.page_digests) == 2
+    assert len(snapshot.page_digests[0]) == 3
+    assert snapshot.page_digests[0][0] != snapshot.page_digests[0][1]
+    assert snapshot.sample_page_digests(max_layers=1, max_pages=2) == (
+        (snapshot.page_digests[0][0], snapshot.page_digests[0][1]),
+    )
+
+
+def test_compare_kv_debug_snapshots_finds_first_mismatch():
+    from sgl_jax.srt.disaggregation.debug_utils import (
+        build_kv_debug_snapshot,
+        count_kv_debug_mismatches,
+        find_first_kv_debug_mismatch,
+    )
+
+    base = np.arange(2 * 2 * 3, dtype=np.int16).reshape(2, 2, 3)
+    modified = base.copy()
+    modified[1, 0, 2] += 7
+
+    left = build_kv_debug_snapshot(base)
+    right = build_kv_debug_snapshot(modified)
+
+    assert count_kv_debug_mismatches(left, right) == 1
+    assert find_first_kv_debug_mismatch(left, right) == (1, 0)
+
+
+def test_build_kv_debug_snapshot_rejects_inputs_without_layer_and_page_axes():
+    from sgl_jax.srt.disaggregation.debug_utils import build_kv_debug_snapshot
+
+    with pytest.raises(ValueError, match="at least 2 dims"):
+        build_kv_debug_snapshot(np.arange(4, dtype=np.int16))
